feat: add relations between User and Restaurant models

+++ routes/modules/restaurant.js
@@ -2,11 +2,12 @@ const express = require('express')
 const router = express.Router()
 const Restaurant = require('../../models/restaurant')
 
-// 瀏覽特定餐廳
+// detail 瀏覽特定餐廳
 router.get("/:restaurantId", (req, res) => {
   // console.log(req.params.restaurantId) // 這資料存在
+  const userId = req.user._id
   const { restaurantId } = req.params
-  Restaurant.findById(restaurantId)
+  return Restaurant.findOne({ restaurantId, userId })
     .lean()
     .then(restaurantData => res.render("show", { restaurantData }))
     .catch(err => console.log(err))
@@ -19,15 +20,19 @@ router.get("/new", (req, res) => {
 
 // 新增餐廳 之 處理頁面
 router.post("/", (req, res) => { //該路由僅用於處理 POST 請求(在新增餐廳表單提交後，將資料新增到資料庫)
-  Restaurant.create(req.body) // 使用 req.body 時，它可以讓你獲取到客戶端在 POST 請求中提交的資料
+  const userId = req.user._id
+  const body = req.body
+  return Restaurant.create({ body, userId })
+    // 使用 req.body 時，它可以讓你獲取到客戶端在 POST 請求中提交的資料
     .then(() => res.redirect("/"))
     .catch(err => console.log(err))
 })
 
-// 編輯餐廳頁面
+// edit 編輯餐廳頁面
 router.get("/:restaurantId/edit", (req, res) => {
+  const userId = req.user._id
   const { restaurantId } = req.params
-  Restaurant.findById(restaurantId)
+  return Restaurant.findOne({ restaurantId, userId })
     .lean()
     .then(restaurantData => res.render("edit", { restaurantData }))
     .catch(err => console.log(err))
@@ -35,12 +40,14 @@ router.get("/:restaurantId/edit", (req, res) => {
 
 // 更新餐廳 （導到 ： 瀏覽特定餐廳）
 router.put("/:restaurantId", (req, res) => {
+  const userId = req.user._id
   const { restaurantId } = req.params
-  Restaurant.findByIdAndUpdate(restaurantId, req.body)
+  const body = req.body
+  return Restaurant.findByIdAndUpdate({ restaurantId, body, userId })
     //可依照專案發展方向自定編輯後的動作，這邊是導向到瀏覽特定餐廳頁面
     .then(() => {
       res.redirect(`/restaurants/${restaurantId}`)
-      console.log(req.body)
+      // console.log(req.body)
     }
     )
     .catch(err => console.log(err))
@@ -48,8 +55,9 @@ router.put("/:restaurantId", (req, res) => {
 
 // 刪除餐廳
 router.delete("/:restaurantId", (req, res) => {
+  const userId = req.user._id
   const { restaurantId } = req.params
-  Restaurant.findByIdAndDelete(restaurantId)
+  return Restaurant.findByIdAndDelete({ restaurantId, userId })
     .then(() => res.redirect("/"))
     .catch(err => console.log(err))
 })
